updating

+++ Personal/install-i3-mint.sh
@@ -1,29 +1,50 @@
 #!/bin/bash
 
 # Install i3 and core dependencies
-sudo apt install i3 i3status rofi dunst xfce4-terminal micro -y
+sudo apt install -y \
+  i3 \
+  i3status \
+  rofi \
+  dunst \
+  xfce4-terminal \
+  micro \
+  polybar \
+  build-essential \
+  cmake \
+  libjsoncpp-dev \
+  libxcb-randr0-dev \
+  libxcb-xinerama0-dev \
+  libxcb-util-dev \
+  libxcb-icccm4-dev \
+  libiw-dev \
+  copyq \
+  blueman \
+  pasystray
 
-# Install Polybar and dependencies
-sudo apt install polybar build-essential cmake libjsoncpp-dev libxcb-randr0-dev libxcb-xinerama0-dev libxcb-util-dev libxcb-icccm4-dev libiw-dev -y
+# Install ProtonVPN if not already installed
+if ! dpkg -l | grep -q protonvpn-stable-release; then
+    wget https://repo.protonvpn.com/debian/dists/stable/main/binary-all/protonvpn-stable-release_1.0.3-3_all.deb -O protonvpn-stable-release.deb
+    sudo apt install --allow-downgrades ./protonvpn-stable-release.deb -y
+    sudo apt update
+    sudo apt install protonvpn -y
+    rm -f protonvpn-stable-release.deb
+else
+    echo "ProtonVPN stable release already installed, skipping."
+fi
 
-# Install autostart apps
-sudo apt install copyq blueman pasystray -y
+# Variables
+USER="brett"
+USER_HOME="/home/$USER"
 
-# Install ProtonVPN
-wget https://protonvpn.com/download/protonvpn-stable-release_1.0.3-3_all.deb
-sudo apt install ./protonvpn-stable-release_1.0.3-3_all.deb -y
-sudo apt update
-sudo apt install protonvpn -y
-
-# Symlink dotfiles (adjust paths as needed)
-mkdir -p ~/.config/{i3,polybar,rofi,dunst}
-ln -s ~/dotfiles/.config/i3/config ~/.config/i3/config
-ln -s ~/dotfiles/.config/polybar ~/.config/polybar
-ln -s ~/dotfiles/.config/rofi/config.rasi ~/.config/rofi/config.rasi
-ln -s ~/dotfiles/.config/dunst/dunstrc ~/.config/dunst/dunstrc
-mkdir -p ~/.bin-personal
-ln -s ~/dotfiles/.bin-personal/* ~/.bin-personal/
-chmod +x ~/.bin-personal/*
+# Symlink dotfiles
+sudo -u "$USER" bash -c "mkdir -p $USER_HOME/.config/i3 $USER_HOME/.config/polybar $USER_HOME/.config/rofi $USER_HOME/.config/dunst"
+sudo -u "$USER" bash -c "ln -sf $USER_HOME/dotfiles/.config/i3/config $USER_HOME/.config/i3/config"
+sudo -u "$USER" bash -c "ln -sf $USER_HOME/dotfiles/.config/polybar $USER_HOME/.config/polybar"
+sudo -u "$USER" bash -c "ln -sf $USER_HOME/dotfiles/.config/rofi/config.rasi $USER_HOME/.config/rofi/config.rasi"
+sudo -u "$USER" bash -c "ln -sf $USER_HOME/dotfiles/.config/dunst/dunstrc $USER_HOME/.config/dunst/dunstrc"
+sudo -u "$USER" bash -c "mkdir -p $USER_HOME/.bin-personal"
+sudo -u "$USER" bash -c "ln -sf $USER_HOME/dotfiles/.bin-personal/* $USER_HOME/.bin-personal/"
+sudo -u "$USER" bash -c "chmod +x $USER_HOME/.bin-personal/*"
 
 # Verify installations
 i3 --version
